+++ lib/src/platform/web/file_picker_web.dart
@@ -5,6 +5,7 @@ import 'dart:typed_data';
 import 'package:file_picker/src/api/file_picker_types.dart';
 import 'package:file_picker/src/api/file_picker_result.dart';
 import 'package:file_picker/src/api/platform_file.dart';
+import 'package:file_picker/src/api/android_saf_options.dart';
 import 'package:file_picker/src/platform/file_picker_platform_interface.dart';
 import 'package:flutter_web_plugins/flutter_web_plugins.dart';
 import 'package:path/path.dart' as p;
@@ -39,6 +40,16 @@ class FilePickerWeb extends FilePickerPlatform {
   }
 
   @override
+  Future<String?> getDirectoryPath({
+    String? dialogTitle,
+    bool lockParentWindow = false,
+    String? initialDirectory,
+    AndroidSAFOptions? androidSafOptions,
+  }) async {
+    throw UnimplementedError('getDirectoryPath() has not been implemented.');
+  }
+
+  @override
   Future<FilePickerResult?> pickFiles({
     String? dialogTitle,
     String? initialDirectory,
@@ -52,6 +63,7 @@ class FilePickerWeb extends FilePickerPlatform {
     bool readSequential = false,
     bool cancelUploadOnWindowBlur = true,
     int compressionQuality = 0,
+    AndroidSAFOptions? androidSafOptions,
   }) async {
     if (type != FileType.custom && (allowedExtensions?.isNotEmpty ?? false)) {
       throw Exception(